requirements.txt
apache-airflow[mysql]
mysqlclient